Splitting the Components
