+++ src/App.js
@@ -1,26 +1,20 @@
-import ExpenseItem from "./components/ExpenseItem";
+// Write your code at relevant places in the code below
+
+import Expenses from "./components/Expenses";
 
 function App() {
+  
   const expenses = [
-    { id: 1, title: "Insurance", date: new Date(2023, 7, 15), price: 100 },
-    { id: 2, title: "Book", date: new Date(2023, 8, 25), price: 10 },
-    { id: 3, title: "Pen", date: new Date(2023, 2, 10), price: 1 },
-    { id: 4, title: "Laptop", date: new Date(2023, 9, 17), price: 200 },
+    { id: "1", date: new Date(2023, 7, 15), title: "Insurance", price: 100 },
+    { id: "2", date: new Date(2023, 3, 25), title: "Book", price: 10 },
+    { id: "3", date: new Date(2023, 10, 11), title: "Pen", price: 1 },
+    { id: "4", date: new Date(2023, 1, 14), title: "Laptop", price: 200 },
   ];
 
   return (
     <div>
-      <h2>Let's get Started</h2>
-      {expenses.map((expense) => {
-        return (
-          <ExpenseItem
-            key={expense.id}
-            title={expense.title}
-            date={expense.date}
-            price={expense.price}
-          ></ExpenseItem>
-        );
-      })}
+      <h1>Let's get Started</h1>
+      <Expenses expenses = {expenses}></Expenses>
     </div>
   );
 }
