+++ src/components/ExpenseItem.js
@@ -1,34 +1,17 @@
+// Write your code at relevant places in the code below
+
 import "./ExpenseItem.css";
-const months = [
-  "January",
-  "February",
-  "March",
-  "April",
-  "May",
-  "June",
-  "July",
-  "August",
-  "September",
-  "October",
-  "November",
-  "December",
-];
+import ExpenseDate from "./ExpenseDate";
 function ExpenseItem(props) {
-  const month = months[props.date.getMonth()];
-  const year = props.date.getFullYear();
-  const date = props.date.getDate().toString().padStart(2, '0');
-  return(
-  <div className="expense-item">
-    <div>
-      <div>{month}</div>
-      <div>{year}</div>
-      <div>{date}</div>
+  return (
+    <div className="expense-item">
+      <ExpenseDate date= {props.date} ></ExpenseDate>
+      <div className="expense-item__description">
+        <h2>{props.title}</h2>
+        <div className="expense-item__price">${props.price}</div>
+      </div>
     </div>
-    <div className="expense-item__description">
-      <h2>{props.title}</h2>
-      <div className="expense-item__price">${props.price}</div>
-    </div>
-  </div>
-  )
+  );
 }
+
 export default ExpenseItem;
